Fix duplicate time indicator bug
