+++ www/src/js/views/timetable/Timetable.jsx
@@ -5,7 +5,7 @@ import { values, flattenDeep, noop } from 'lodash';
 import type { Lesson } from 'types/modules';
 import type { TimetableArrangement } from 'types/timetables';
 
-import { SCHOOLDAYS, calculateBorderTimings } from 'utils/timify';
+import { SCHOOLDAYS, calculateBorderTimings, getCurrentDayIndex } from 'utils/timify';
 
 import styles from './Timetable.scss';
 import TimetableTimings from './TimetableTimings';
@@ -32,7 +32,7 @@ class Timetable extends PureComponent<Props> {
 
     const lessons: Array<Lesson> = flattenDeep(values(this.props.lessons));
     const { startingIndex, endingIndex } = calculateBorderTimings(lessons);
-    const currentDayIndex: number = new Date().getDay() - 1; // Monday = 0, Friday = 4
+    const currentDayIndex: number = getCurrentDayIndex(); // Monday = 0, Friday = 4
 
     return (
       <div
