+++ www/src/js/views/timetable/TimetableDay.scss
@@ -86,3 +86,7 @@
   padding: 0 0.1rem 0 0;
   background: linear-gradient(to bottom, var(--gray-lightest) 50%, transparent 50%);
 }
+
+.currentDay {
+  box-shadow: inset 0 0 0 99999px rgba(128, 128, 128, 0.1);
+}
