+++ www/src/js/views/timetable/_variables.scss
@@ -9,3 +9,8 @@ $timetable-day-v-size: 2.5rem;
 
 $timetable-row-height: 3.6rem;
 $timetable-col-width: 4rem;
+
+$current-time-indicator-circle-diameter: 0.5rem;
+$current-time-indicator-circle-offset: -0.25rem;
+$current-time-indicator-line-width: 0.1rem;
+$current-time-indicator-line-offset: -0.05rem;
